rename meteo_decorator to decorators
